Moved another example from Main to Examples
