+++ de.hamstersimulator.objectsfirst.examples/src/main/java/module-info.java
@@ -2,4 +2,6 @@ module de.hamstersimulator.objectsfirst.examples {
     requires de.hamstersimulator.objectsfirst.main;
     requires de.hamstersimulator.objectsfirst.ui;
     requires de.hamstersimulator.objectsfirst.inspector;
+    
+    requires javafx.graphics;
 }
